requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
boto3==1.17.69
botocore==1.20.69
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-extensions==3.1.3
django-formtools==2.2
django-storages==1.11.1
djangorestframework==3.12.4
future==0.18.2
idna==2.10
jmespath==0.10.0
musicbrainzngs==0.7.1
pidi==0.0.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pydotplus==2.0.2
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-mpd2==3.0.4
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
